requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
torch==2.5.0
torchvision==0.20.0
Pillow==10.3.0
gunicorn==21.2.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
pymongo==4.11.1